refactor: resolve `package.json` files directly
`readPackageJSON` is generally used to traverse the filesystem for the
closest `package.json`, then attempt to parse it via `JSON.parse`.

In all cases, we pass an absolute path to a `package.json`, which means
we don't actually make use of what this function does (other than
`JSON.parse` at the very end).

This refactors such that we simply `JSON.parse` ourselves and use the
`PackageJson` type from `pkg-types`.

We also currently re-read the `package.json` when we want to
`writeDeps`. This is so we can revert our changes if needed. Since we
already read the file earlier in the stack, we now pass that down
directly.